moved component to OmouComponents, listed height as required prop, and renamed rest prop more descriptively
